Godfile data-store step 1: extract sqlite-migrations
Move the SQLite migration routines out of packages/data-store/src/sqlite-adapter.ts
into a new packages/data-store/src/sqlite-migrations.ts module, plus a shared
sqlite-executor.ts context and a pure sqlite-external-dependencies.ts helper.

The adapter keeps the same public methods; each now delegates to a free function
over a narrow SqliteExecutor context.

The moved catch blocks previously swallowed every caught error silently, behind
"table may not exist yet" comments that are unreachable now that initSchema() runs
before migrate() in the constructor. They now log through console.warn with the
failing function and the row/context, so a real SQL or JSON failure is debuggable
instead of vanishing. Control flow is preserved: migrations still tolerate missing
tables and malformed rows and never abort startup. The COLUMN_MIGRATIONS loop keeps
rethrowing anything other than the idempotent "duplicate column name".

migrateTaskExternalDependenciesToWorkflows also stops silently dropping data: the
old bulk UPDATE cleared external_dependencies on every task, including malformed
rows and orphaned tasks whose workflow_id has no matching workflow, even though
their deps were never promoted. It now clears only the tasks whose deps were
merged into an existing workflow, leaving malformed and orphaned rows intact.

Review claim: migrate* logic moves from sqlite-adapter.ts to sqlite-migrations.ts as
free functions over a SqliteExecutor context, adapter methods become one-line
delegates, every caught migration error is now logged with context instead of
being silently swallowed, and the workflow-dependency migration clears only the
task rows it actually promoted.
Review lane: behavior
Review unit: ownership-refactor
